Move LoadSchema into its own component and use its own slice details
