+++ app/controllers/concerns/transformation_redux_state.rb
@@ -8,7 +8,7 @@ module TransformationReduxState
       entities: {
         fields: fields_slice, rawRecord: raw_record_slice,
         appDetails: app_details_slice, sharedDefinitions: shared_definitions_slice, 
-        schemas: schema_slice, schema_fields: schema_fields_slice
+        schemas: schema_slice, schemaFields: schema_fields_slice
       },
       ui: {
         fields: ui_fields_slice, appDetails: ui_app_details_slice
